feat: add check format vehicle type data

+++ backend/core/vehicle_type.py
@@ -1,3 +1,13 @@
+from backend.exceptions.vehicle_type_exceptions import VehicleTypeUpdateException
+from backend.exceptions.vehicle_type_exceptions import VehicleTypeGetAllException
+from backend.exceptions.vehicle_type_exceptions import VehicleTypeGetCategoryException
+from backend.exceptions.vehicle_type_exceptions import VehicleTypeFormatDataException
+from backend.exceptions.vehicle_type_exceptions import VehicleTypeCreateException
+from backend.exceptions.vehicle_type_exceptions import VehicleTypeDeleteException
+from backend.service.check_vehicle_type_data import CheckVehicleTypeDataFormat
+from backend.service.convert_uppercase import convert_uppercase
+
+
 class VehicleType:
     """Class Vehicle Type."""
 
@@ -9,46 +19,61 @@ class VehicleType:
 
     def get_all(self):
         """Get all vehicles type."""
-        self.sql = "SELECT * FROM VEHICLES_TYPE"
-        self.cursor.execute(self.sql)
+        try:
+            self.sql = "SELECT * FROM VEHICLES_TYPE"
+            self.cursor.execute(self.sql)
+            values = list(self.cursor.fetchall())
+        except Exception:
+            raise VehicleTypeGetAllException()
         columns = list(map(lambda x: x[0], self.cursor.description))
-        values = list(self.cursor.fetchall())
         vehicles_types = [dict(zip(columns, value)) for value in values]
         return vehicles_types
 
-    def get_by_id(self, vehicle_type_id:int):
+    def get_by_id(self, vehicle_type_id: int):
         """Get vehicle type by vehicle_type_id."""
-        self.sql = "SELECT * FROM VEHICLES_TYPE WHERE vehicle_type_id=?"
-        values = self.cursor.execute(self.sql, (vehicle_type_id,))
+        try:
+            self.sql = "SELECT * FROM VEHICLES_TYPE WHERE vehicle_type_id=?"
+            values = self.cursor.execute(self.sql, (vehicle_type_id,))
+        except Exception:
+            raise VehicleTypeGetCategoryException()
         columns = list(map(lambda x: x[0], self.cursor.description))
         vehicles_type = [dict(zip(columns, value)) for value in values]
         return vehicles_type
 
-    def get_by_brand(self, brand:str):
+    def get_by_brand(self, brand: str):
         """Get vehicle type by brand."""
-        self.sql = "SELECT * FROM VEHICLES_TYPE WHERE brand=?"
-        values = self.cursor.execute(self.sql, (brand,))
+        try:
+            self.sql = "SELECT * FROM VEHICLES_TYPE WHERE brand=?"
+            values = self.cursor.execute(self.sql, (brand,))
+        except Exception:
+            raise VehicleTypeGetCategoryException()
         columns = list(map(lambda x: x[0], self.cursor.description))
         vehicle_types = [dict(zip(columns, value)) for value in values]
         return vehicle_types
 
-    def get_by_model(self, model:str):
+    def get_by_model(self, model: str):
         """Get vehicle type by model."""
-        self.sql = "SELECT * FROM VEHICLES_TYPE WHERE model=?"
-        values = self.cursor.execute(self.sql, (model,))
+        try:
+            self.sql = "SELECT * FROM VEHICLES_TYPE WHERE model=?"
+            values = self.cursor.execute(self.sql, (model,))
+        except Exception:
+            raise VehicleTypeGetCategoryException()
         columns = list(map(lambda x: x[0], self.cursor.description))
         vehicle_types = [dict(zip(columns, value)) for value in values]
         return vehicle_types
 
-    def get_by_year(self, year:int):
+    def get_by_year(self, year: int):
         """Get vehicle type by year."""
-        self.sql = "SELECT * FROM VEHICLES_TYPE WHERE year=?"
-        values = self.cursor.execute(self.sql, (year,))
+        try:
+            self.sql = "SELECT * FROM VEHICLES_TYPE WHERE year=?"
+            values = self.cursor.execute(self.sql, (year,))
+        except Exception:
+            raise VehicleTypeGetCategoryException()
         columns = list(map(lambda x: x[0], self.cursor.description))
         vehicle_types = [dict(zip(columns, value)) for value in values]
         return vehicle_types
 
-    def get_vehicle_type_id(self, brand:str, model: str, year:int):
+    def get_vehicle_type_id(self, brand: str, model: str, year: int):
         """Get vehicle type id by the brand, model and year."""
         vehicle_type_id = None
         vehicle_types = self.get_all()
@@ -59,43 +84,67 @@ class VehicleType:
                         vehicle_type_id = vehicle_type.get("vehicle_type_id")
         return vehicle_type_id
 
-    def create(self, data:dict):
-        """Create a vehicle type."""
-        columns = ",".join([*data.keys()])
-        values = list(data.values())
-        placeholders = ','.join(['?'] * len(data))
+    def create(self, **kwargs):
+        """Create a vehicle type
+
+         **kwargs:
+        brand: vehicle brand
+        model: vehicle model
+        year: vehicle year.
+        """
+
+        try:
+            CheckVehicleTypeDataFormat(vehicle_type_data=kwargs)
+        except VehicleTypeFormatDataException as error:
+            raise VehicleTypeFormatDataException(message=error.message)
+        kwargs = convert_uppercase(param=kwargs)
+        columns = ",".join([*kwargs.keys()])
+        values = list(kwargs.values())
+        placeholders = ','.join(['?'] * len(kwargs))
         try:
             self.sql = f"INSERT INTO VEHICLES_TYPE({columns}) VALUES({placeholders})"
             self.cursor.execute(self.sql, values)
             self.connection.commit()
-            result = True
         except Exception:
-            result = False
-        return result
+            raise VehicleTypeCreateException()
 
-    def update(self, vehicle_type_id: int, data:dict):
-        """Update vehicle type details."""
-        values = list(data.values())
+    def update(self, vehicle_type_id: int, **kwargs):
+        """Update vehicle type details.
+
+         **kwargs:
+        brand: vehicle brand
+        model: vehicle model
+        year: vehicle year.
+        """
+
+        try:
+            CheckVehicleTypeDataFormat(vehicle_type_data=kwargs)
+        except VehicleTypeFormatDataException as error:
+            raise VehicleTypeFormatDataException(message=error.message)
+        kwargs = convert_uppercase(param=kwargs)
+        values = list(kwargs.values())
         values.append(vehicle_type_id)
-        columns = list(data.keys())
+        columns = list(kwargs.keys())
         columns = [column + " = ?" for column in columns]
         columns = " , ".join(columns)
         try:
             self.sql = f"UPDATE VEHICLES_TYPE SET {columns} WHERE vehicle_type_id = ?"
-            self.cursor.execute(self.sql, (values))
+            self.cursor.execute(self.sql, values)
             self.connection.commit()
-            result = True
         except Exception:
-            result = False
-        return result
+            raise VehicleTypeUpdateException()
 
-    def delete(self, vehicle_type_id:int):
+    def delete(self, vehicle_type_id: int):
         """Delete a vehicle type."""
+        self.sql = "SELECT * FROM VEHICLES WHERE vehicle_type_id=?"
+        values = self.cursor.execute(self.sql, (vehicle_type_id,))
+        response = [value for value in values]
+        if response:
+            error_message = "The vehicle type contains a vehicle registered."
+            raise VehicleTypeDeleteException(message=error_message)
         try:
             self.sql = "DELETE FROM VEHICLES_TYPE WHERE vehicle_type_id=?"
             self.cursor.execute(self.sql, (vehicle_type_id,))
             self.connection.commit()
-            result = True
         except Exception:
-            result = False
-        return result
+            raise VehicleTypeDeleteException()
